+++ .vs/Sparta-RPG2-/v17/DocumentLayout.json
@@ -3,8 +3,20 @@
   "WorkspaceRootPath": "C:\\WorkPlace\\Sparta-RPG2-\\",
   "Documents": [
     {
+      "AbsoluteMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|c:\\workplace\\sparta-rpg2-\\sparta-rpg2-\\inventory.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|solutionrelative:sparta-rpg2-\\inventory.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|c:\\workplace\\sparta-rpg2-\\sparta-rpg2-\\itemequipped.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|solutionrelative:sparta-rpg2-\\itemequipped.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
       "AbsoluteMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|c:\\workplace\\sparta-rpg2-\\sparta-rpg2-\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
       "RelativeMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|solutionrelative:sparta-rpg2-\\program.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
+    },
+    {
+      "AbsoluteMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|c:\\workplace\\sparta-rpg2-\\sparta-rpg2-\\character.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}",
+      "RelativeMoniker": "D:0:0:{68E71864-37C0-434F-B2DC-774AE797B7DB}|Sparta-RPG2-\\Sparta-RPG2-.csproj|solutionrelative:sparta-rpg2-\\character.cs||{A6C744A8-0E4A-4FC6-886A-064283054674}"
     }
   ],
   "DocumentGroupContainers": [
@@ -14,17 +26,56 @@
       "DocumentGroups": [
         {
           "DockedWidth": 200,
-          "SelectedChildIndex": 0,
+          "SelectedChildIndex": 2,
           "Children": [
             {
               "$type": "Document",
+              "DocumentIndex": 3,
+              "Title": "Character.cs",
+              "DocumentMoniker": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Character.cs",
+              "RelativeDocumentMoniker": "Sparta-RPG2-\\Character.cs",
+              "ToolTip": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Character.cs",
+              "RelativeToolTip": "Sparta-RPG2-\\Character.cs",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAUAAAAVAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-04-21T10:20:17.259Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 1,
+              "Title": "ItemEquipped.cs",
+              "DocumentMoniker": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\ItemEquipped.cs",
+              "RelativeDocumentMoniker": "Sparta-RPG2-\\ItemEquipped.cs",
+              "ToolTip": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\ItemEquipped.cs",
+              "RelativeToolTip": "Sparta-RPG2-\\ItemEquipped.cs",
+              "ViewState": "AgIAABAAAAAAAAAAAAAIwCkAAAAOAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-04-21T10:18:18.424Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
               "DocumentIndex": 0,
+              "Title": "Inventory.cs",
+              "DocumentMoniker": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Inventory.cs",
+              "RelativeDocumentMoniker": "Sparta-RPG2-\\Inventory.cs",
+              "ToolTip": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Inventory.cs",
+              "RelativeToolTip": "Sparta-RPG2-\\Inventory.cs",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAcAAAAeAAAAAAAAAA==",
+              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
+              "WhenOpened": "2025-04-21T10:15:58.308Z",
+              "EditorCaption": ""
+            },
+            {
+              "$type": "Document",
+              "DocumentIndex": 2,
               "Title": "Program.cs",
               "DocumentMoniker": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Program.cs",
               "RelativeDocumentMoniker": "Sparta-RPG2-\\Program.cs",
               "ToolTip": "C:\\WorkPlace\\Sparta-RPG2-\\Sparta-RPG2-\\Program.cs",
               "RelativeToolTip": "Sparta-RPG2-\\Program.cs",
-              "ViewState": "AgIAAAAAAAAAAAAAAAAAAAkAAAAnAAAAAAAAAA==",
+              "ViewState": "AgIAABIAAAAAAAAAAAAIwBAAAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
               "WhenOpened": "2025-04-21T08:51:39.662Z",
               "EditorCaption": ""
